continue simplifying amwg
