+++ src/python/packages/amwg/amwg2.py
@@ -3,7 +3,8 @@
 # DIAG Set 2 - Line Plots of Annual Implied Northward Transport
 
 from pprint import pprint
-from metrics.packages.amwg.amwg import amwg_plot_plan, src2modobs, src2obsmod
+from metrics.packages.amwg.amwg import amwg_plot_plan
+from metrics.packages.amwg.tools import src2modobs, src2obsmod
 from metrics.packages.amwg.derivations.vertical import *
 from metrics.packages.plotplan import plot_plan
 from metrics.computation.reductions import *
