+++ src/python/packages/amwg/amwg3.py
@@ -3,7 +3,8 @@
 # DIAG Set 3 - Line Plots of  Zonal Means
 
 from pprint import pprint
-from metrics.packages.amwg.amwg import amwg_plot_plan, src2modobs, src2obsmod
+from metrics.packages.amwg.amwg import amwg_plot_plan
+from metrics.packages.amwg.tools import src2modobs, src2obsmod
 from metrics.packages.amwg.amwg5and6 import amwg_plot_set5and6
 from metrics.packages.amwg.derivations.vertical import *
 from metrics.packages.plotplan import plot_plan
