+++ src/python/packages/amwg/amwg4.py
@@ -4,7 +4,8 @@
 # DIAG Set $A - Horizontal Contour Plots of Meridional Means
 
 from pprint import pprint
-from metrics.packages.amwg.amwg import amwg_plot_plan, src2modobs, src2obsmod
+from metrics.packages.amwg.amwg import amwg_plot_plan
+from metrics.packages.amwg.tools import src2modobs, src2obsmod
 from metrics.packages.amwg.amwg1 import reduced_variables_press_lev, reduced_variables_hybrid_lev
 from metrics.packages.amwg.derivations.vertical import *
 from metrics.packages.plotplan import plot_plan
