+++ src/python/packages/amwg/amwg9.py
@@ -3,7 +3,8 @@
 # DIAG Set 9 - Horizontal Contour Plots of DJF-JJA Differences
 
 from pprint import pprint
-from metrics.packages.amwg.amwg import amwg_plot_plan, src2modobs, src2obsmod
+from metrics.packages.amwg.amwg import amwg_plot_plan
+from metrics.packages.amwg.tools import src2modobs, src2obsmod
 from metrics.packages.amwg.derivations.vertical import *
 from metrics.packages.plotplan import plot_plan
 from metrics.computation.reductions import *
